feat: use custom mapping for settings
